requirements.txt
# numpy
# opencv-contrib-python
opencv-python
openpyxl
# pandas
# pillow
pyttsx3
opencv-contrib-python==4.8.1.78
pandas>=1.5.0
numpy>=1.21.0
Pillow>=9.0.0